Pack JavaFX application into seperate class
